Option 2leg (mleg submit to sell)   example
As discussed in the buying 2leg (mleg) verticle spread, the limitations for windows and python SDK makes it so that you have to use a two approach for it, using CURL's to submit your sell order in.   As you can use this on your paper account and successfully sell them.